microservices/cart/pyproject.toml
[tool.poetry]
name = "cart"
version = "0.1.0"
description = ""
authors = ["Clariteia Devs <devs@clariteia.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.3.2"
minos-microservice-common = { git = "https://github.com/Clariteia/minos_microservice_common.git", rev="624e9a35b3a0a464f819d2b1e1bfd053ec3942ca" }
minos-microservice-networks = "^0.0.12"
minos-microservice-saga = "^0.0.9"
minos-microservice-cqrs = "^0.0.1"
SQLAlchemy = "^1.4.22"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
microservice = "src.cli:main"